Migrate Java 11 tests out of supersource
